requirements.txt
click
jinja2
pillow

